iOS examples updated to allow targeting of arm64 architecture, as well as armv7. Buddy: Ian

+++ ios/ios_src/AppLocationDelegate.mm
@@ -58,7 +58,7 @@ ViewController* m_pUIViewController;
 {
     if (newHeading.headingAccuracy >= 0)
     {
-        float heading = newHeading.trueHeading;
+        float heading = static_cast<float>(newHeading.trueHeading);
         
         if (m_pUIViewController.interfaceOrientation == UIInterfaceOrientationLandscapeLeft)
         {
